+++ tests/programmability.py
@@ -652,6 +652,7 @@ if __name__ == "__main__":
         package="samples/apps/logging/logging",
         nodes=infra.e2e_args.min_nodes(cr.args, f=1),
         jwt_key_refresh_interval_s=100000,
+        jwt_key_refresh_max_response_size="4KB",
         issuer_port=12346,
         forwarding_timeout_ms=jwt_forwarding_timeout_ms,
     )